æ'virker

+++ src/gui/model/SongListModel.java
@@ -58,7 +58,7 @@ public class SongListModel {
     }
 
 
-    public void createSong(String title, String artist, String genre, int duration, String pathToFile) throws DALException {
+    public void createSong(String title, String artist, String genre, double duration, String pathToFile) throws DALException {
         songsToBeViewed.add(new SongModel(songManager.createSong(title, artist, genre, duration, pathToFile)));
     }
